requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
click==8.1.7
pydantic>=2.7.4,<3.0.0
python-dotenv==1.0.0
httpx==0.25.2

# LangGraph Dependencies
langgraph>=0.2.34
langchain-core>=0.3.72

# Document Processing
PyPDF2==3.0.1
python-docx==0.8.11
langchain==0.3.27
langchain-community==0.3.27
python-multipart==0.0.6
unstructured==0.11.6

# Vector Database
pymilvus==2.3.4
milvus==2.3.4

# Text Processing
jieba==0.42.1
scikit-learn==1.3.2
numpy>=1.26.2

# Logging and Utils
loguru==0.7.2
pyyaml==6.0.1
tqdm==4.66.1

# Development Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1